+++ lib/flack/app/executions.rb
@@ -25,7 +25,11 @@ class Flack::App
       #
     q = q.order(:exid)
 
-    respond(env, q.all)
+    if query_value(env, 'count')
+      respond(env, { count: q.count })
+    else
+      respond(env, q.all)
+    end
   end
 
   # GET /executions/<id>